% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conservedMotifs.R
\name{conservedMotifs}
\alias{conservedMotifs}
\title{check the conserved motifs in the orthologs}
\usage{
conservedMotifs(aln, aln_hs, aln_mm, PWMs, queryGenome, background = "genome")
}
\arguments{
\item{aln}{alignment of multiple DNAs. Output of \link{alignment} function.}

\item{aln_hs, aln_mm}{output of link{searchTFBPS} for human and mouse.}

\item{PWMs}{The Position Weight Matrix list represented as a numeric matrix.
Object of \link[TFBSTools:XMatrixList]{PWMatrixList} or
\link[TFBSTools:XMatrixList]{PFMatrixList}.}

\item{queryGenome}{An object of \link[BSgenome:BSgenome-class]{BSgenome} for
query enhancer.}

\item{background}{background nucleotide frequencies. Default is "genome".
Refer \link[motifmatchr]{matchMotifs} for details.}
}
\value{
A list of XStringviews.
}
\description{
Print the conserved motifs in the alignments
}
\examples{
library(BSgenome.Hsapiens.UCSC.hg38)
library(BSgenome.Mmusculus.UCSC.mm10)
library(BSgenome.Drerio.UCSC.danRer10)
LEN <- GRanges("chr4", IRanges(19050041, 19051709))
seqEN <- getSeq(BSgenome.Drerio.UCSC.danRer10, LEN)
aln_hs <- readRDS(system.file("extdata", "aln_hs.rds",
               package="enhancerHomologSearch"))
genome(aln_hs) <- Hsapiens
aln_mm <- readRDS(system.file("extdata", "aln_mm.rds",
               package="enhancerHomologSearch"))
genome(aln_mm) <- Mmusculus
al <- alignment(seqEN, list(human=aln_hs, mouse=aln_mm),
                method="ClustalW", order="input")
data(motifs)
conservedMotifs(al[[1]], aln_hs, aln_mm, motifs[["dist60"]], Drerio)
}
