% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancers.R
\docType{class}
\name{Enhancers-class}
\alias{Enhancers-class}
\alias{Enhancers}
\alias{$,Enhancers-method}
\alias{$<-,Enhancers-method}
\alias{distance,Enhancers,ANY-method}
\alias{distance}
\alias{distance,Enhancers-method}
\alias{distance<-,Enhancers-method}
\alias{distance<-}
\alias{distance<-,Enhancers,ANY-method}
\alias{tfbp,Enhancers-method}
\alias{tfbp}
\alias{tfbp,Enhancers,ANY-method}
\alias{query_tfbp,Enhancers-method}
\alias{query_tfbp}
\alias{query_tfbp,Enhancers,ANY-method}
\alias{getSeq,Enhancers-method}
\alias{subsetByOverlaps,Enhancers,ANY-method}
\alias{subsetByOverlpas,Enhancers-method}
\alias{subset,Enhancers-method}
\alias{seqinfo,Enhancers-method}
\alias{genome,Enhancers-method}
\alias{genome}
\alias{genome<-,Enhancers-method}
\alias{genome<-}
\alias{genome<-,Enhancers,BSgenome-method}
\alias{peaks,Enhancers-method}
\alias{peaks}
\alias{peaks<-,Enhancers-method}
\alias{peaks<-}
\alias{peaks<-,Enhancers,GRanges-method}
\alias{coerce}
\alias{coerce,Enhancers,GRanges-method}
\alias{show,Enhancers-method}
\title{Class \code{"Enhancers"}}
\usage{
Enhancers(genome, peaks, TFBP, TFBP0)

\S4method{$}{Enhancers}(x, name)

\S4method{$}{Enhancers}(x, name) <- value

\S4method{distance}{Enhancers,ANY}(x)

\S4method{distance}{Enhancers}(x) <- value

\S4method{tfbp}{Enhancers}(x)

\S4method{query_tfbp}{Enhancers}(x)

\S4method{getSeq}{Enhancers}(x, ...)

\S4method{subsetByOverlaps}{Enhancers,ANY}(
  x,
  ranges,
  maxgap = -1L,
  minoverlap = 0L,
  type = c("any", "start", "end", "within", "equal"),
  invert = FALSE,
  ...
)

\S4method{subset}{Enhancers}(x, ...)

\S4method{seqinfo}{Enhancers}(x)

\S4method{genome}{Enhancers}(x)

\S4method{genome}{Enhancers}(x) <- value

\S4method{peaks}{Enhancers}(x)

\S4method{peaks}{Enhancers}(x) <- value

\S4method{show}{Enhancers}(object)
}
\arguments{
\item{genome}{An object of \link[BSgenome:BSgenome-class]{BSgenome}.}

\item{peaks}{An object of \link[GenomicRanges:GRanges-class]{GRanges}.}

\item{TFBP}{An object of \link[Matrix:lsparseMatrix-class]{lgCMatrix}.}

\item{TFBP0}{An vector of logical.
\code{"Enhancers"}}

\item{x}{An object of \code{"Enhancers"}}

\item{name}{Slot name.}

\item{value}{The values.}

\item{\dots}{parameters can be passed to upstream functions.}

\item{ranges, maxgap, minoverlap, type, invert}{parameters used by
\link[GenomicRanges:findOverlaps-methods]{subsetByOverlaps}}

\item{object}{An object of \code{"Enhancers"}}
}
\value{
An object of Enhancers.
}
\description{
An object of class "Enhancers"
represents the output of function \link{getENCODEdata},
which includes the sequences of enhancers and their genomic coordinates.
}
\section{Slots}{

\describe{
\item{\code{genome}}{An object of \link[BSgenome:BSgenome-class]{BSgenome}.}

\item{\code{peaks}}{An object of \link[GenomicRanges:GRanges-class]{GRanges}.}

\item{\code{TFBP}}{An object of \link[Matrix:lsparseMatrix-class]{lgCMatrix}.}

\item{\code{TFBP0}}{An vector of logical.}
}}

\examples{
Enhancers()

}
