% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle.R
\name{shuffle}
\alias{shuffle}
\title{shuffle reads}
\usage{
shuffle(reads, k = 2, n = 2)
}
\arguments{
\item{reads}{An object of \link[Biostrings:XStringSet-class]{BStringSet}.}

\item{k}{the k-let size.}

\item{n}{the number of random sequences to generate.}
}
\value{
An object of \link[Biostrings:XStringSet-class]{BStringSet}.
}
\description{
Uses the uShuffle library to shuffle reads
}
\examples{
library(Biostrings)
f <- DNAStringSet(c("CTC-NACCAGTAT", "TTGA", "TACCTAGAG"))
shuffle(f)
}
\references{
Jiang, M., Anderson, J., Gillespie, J. et al.
uShuffle: A useful tool for shuffling biological sequences while preserving
 the k-let counts. BMC Bioinformatics 9, 192 (2008).
 https://doi.org/10.1186/1471-2105-9-192
}
