% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{alignment}
\alias{alignment}
\title{Output}
\usage{
alignment(query, subject, ...)
}
\arguments{
\item{query}{An object of DNAStringSet to represent enhancer}

\item{subject}{An list of objects of \link{enhancers}.}

\item{\dots}{Parameters to be used by \link[msa:msa]{msa}.}
}
\value{
An object of \link{enhancers}.
}
\description{
Do pairwise alignment for query enhancer to target genome
}
\examples{
library(BSgenome.Hsapiens.UCSC.hg38)
library(BSgenome.Mmusculus.UCSC.mm10)
library(BSgenome.Drerio.UCSC.danRer10)
hbegfEN <- GRanges("chr14", IRanges(6760805,	6761115))
seqEN <- getSeq(BSgenome.Drerio.UCSC.danRer10, hbegfEN)
aln_hs <- readRDS(system.file("extdata", "aln_hs.rds",
               package="enhancerHomologSearch"))
aln_mm <- readRDS(system.file("extdata", "aln_mm.rds",
               package="enhancerHomologSearch"))
al <- alignment(seqEN, list(human=aln_hs, mouse=aln_mm),
                method="ClustalOmega", order="input")
}
