% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancers.R
\docType{class}
\name{enhancers-class}
\alias{enhancers-class}
\alias{enhancers}
\alias{$,enhancers-method}
\alias{$,enhancer-method}
\alias{$<-,enhancer-method}
\alias{$<-,enhancers-method}
\alias{getSeq,enhancers-method}
\alias{subsetByOverlaps,enhancers,ANY-method}
\alias{subsetByOverlpas,enhancers-method}
\alias{subset,enhancers-method}
\alias{genome}
\alias{genome,enhancers-method}
\alias{coerce}
\alias{coerce,enhancers,GRanges-method}
\alias{show,enhancers-method}
\title{Class \code{"enhancers"}}
\usage{
enhancers(...)

\S4method{$}{enhancers}(x, name)

\S4method{$}{enhancers}(x, name) <- value

\S4method{getSeq}{enhancers}(x, ...)

\S4method{subsetByOverlaps}{enhancers,ANY}(
  x,
  ranges,
  maxgap = -1L,
  minoverlap = 0L,
  type = c("any", "start", "end", "within", "equal"),
  invert = FALSE,
  ...
)

\S4method{subset}{enhancers}(x, ...)

\S4method{show}{enhancers}(object)
}
\arguments{
\item{\dots}{Each argument in \dots becomes an slot in the new("enhancers")
or An object of \code{"GRanges"} for getSeq,enhancers-method
\code{"enhancers"}}

\item{x}{An object of \code{"enhancers"}}

\item{name}{Slot name.}

\item{value}{The values.}

\item{ranges, maxgap, minoverlap, type, invert}{parameters used by
\link[GenomicRanges:findOverlaps-methods]{subsetByOverlaps}}

\item{object}{An object of \code{"enhancers"}}
}
\value{
An object of enhancers.
}
\description{
An object of class "enhancers"
represents the output of function \link{getENCODEdata},
which includes the sequences of enhancers and their genomic coordinates.
}
\section{Slots}{

\describe{
\item{\code{genome}}{An object of \link[BSgenome:BSgenome-class]{BSgenome}.}

\item{\code{peaks}}{An object of \link[GenomicRanges:GRanges-class]{GRanges}.}
}}

\examples{
enhancers()

}
