% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveAlignments.R
\name{saveAlignments}
\alias{saveAlignments}
\title{output alignments}
\usage{
saveAlignments(al, output_folder = ".")
}
\arguments{
\item{al}{output of \link{alignment}.}

\item{output_folder}{output folder.}
}
\description{
Save enhancer homologs to file in phylip format.
}
\examples{
al <- readRDS(system.file("extdata", "al.rds",
               package="enhancerHomologSearch"))
tmpfolder <- tempdir()
saveAlignments(al, output_folder=tmpfolder)
}
